requirements.txt
streamlit
fastapi
uvicorn
groq
python-dotenv
requests
pydantic
